requirements.txt
azure-functions
pandas
snowflake-connector-python
python-dotenv
